10821: Use ResponseDestroyer in mission destroy

+++ app/destroyers/enumerating_destroyer.rb
@@ -10,7 +10,7 @@ class EnumeratingDestroyer < ApplicationDestroyer
     scope.each do |object|
       next if handle_explicit_skip(object)
       begin
-        raise ActiveRecord::DeleteRestrictionError unless ability.can?(:destroy, object)
+        raise ActiveRecord::DeleteRestrictionError unless ability.nil? || ability.can?(:destroy, object)
         object.destroy
         counts[:destroyed] += 1
       # The former is deprecated, and we are moving toward the latter.
